moved mutation and query from resolvers
