+++ starters/deno-oak-denodb/src/db/resolvers/mutationHandler.ts
@@ -1,0 +1,38 @@
+import { Technologies } from '../model/technology.ts';
+export interface ITechnology {
+	displayName: string;
+	description: string;
+	url: string;
+}
+
+export const createTechnology = async (
+	_: unknown,
+	{ technology: { displayName, description, url } }: any,
+) => {
+	const createdTechnology = await Technologies.create({
+		id: crypto.randomUUID(),
+		displayName,
+		description,
+		url,
+	});
+	return createdTechnology;
+};
+
+export const updateTechnology = async (
+	_: unknown,
+	{ id, value }: { id: string; value: ITechnology },
+) => {
+	await Technologies.where('id', id).update({
+		...value,
+	});
+	return {
+		done: true,
+	};
+};
+
+export const deleteTechnologyById = async (_: unknown, { id }: any) => {
+	await Technologies.deleteById(id);
+	return {
+		done: true,
+	};
+};
